+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to
 
 ### Changed
 
+## [0.4.0]
+
 - Renamed `derate_covariance` parameter `accuracy` to `precision`.
 - Use Generalised Chi Squared distribution to calculate critical values.